Modifies the way how riders are stored inside of RideOut document in Mongo.
Only the IDs of users are stored now and the RideOut DAO now uses
aggrogation for joining the data. Also a SimpleUser class is introduced
which only contains id, username, first and last name.
RideOut class now stores collection of simple users which limits the
amount of information send from back-end to front-end.

I have also modified configuration of the ObjectMapper which Jackson uses to
parse to/from JSON. With the new configuration it will not include
fields with null value in parsed JSON.